attempt to implement suggestions;
