+++ src/ase/cw/model/OrderProducerListener.java
@@ -1,0 +1,8 @@
+package ase.cw.model;
+import java.util.concurrent.BlockingQueue;
+
+public interface OrderProducerListener {
+
+    void onOrderProduced(BlockingQueue order, Order producedOrder);
+
+}
